Added user data creation, as well as fetching
A couple things added here

1. Added a cloudfunction that triggers when a new user is created. This function populates firestore with an entry for the new user.

2. Now that firestore has the new user we can fetch that user when auth is established. This lets us load the podcast page.

BUG ALERT:

Currently if you create a new user the function to create that user in the db is asyncrnously triggered. However the current code
attempts to fetch a new user on AuthStateChanged. This means on initial creation a user may not yet be avalible, causeing an error.
The user data fetching logic should be moved out of the AuthStateChanged. I think it should just be a pub/sub type listener tht updates
the user in realtime.