Adding initial page for creating a podcast
This is the precreation page. It requires the user to enter a podcast title.
At this point clicking continue  does nothing.

The expeccted behavior for continue will be to create a podcast in draft mode on the backend.
This will have a database entry that a user can revisit to make changes, and add metadata to.