Changing folder structure from nested /root/front-end/src -> /root/src
Originall we thought of structuring this as a monorepo. With the front-end and back-end application being rooted at seperate folders.
However it turned out to be more convienent to just root the backend function in the same directory where firebase was initialized (front-end).

Additionally i found that our CI build in github actions was failing due to it expecting a `package.json` in the root location. Overall its easier to just restructure the folder layout like this.