Got rid of the user creation function
Creating the user via function sounded like a good idea, and while it was blazing fast in testing, in production it lead to a broken experience.
Instead I have rewritten the logic to simply check for the existence of a user on sign in, and if the user does not exist we make one.

This is a much easier to unederstand experience.